Convert testing to Conda Env instead of Conda Build
This PR moves the testing environment from relying on Conda-build
and instead creates conda envs which should speed up and simplify tests.

The trade off is the maintenance of multiple files for dependencies
going forward, but I think the benefits outweigh the costs.

Borrows from the QCFractal script from @dgasmith to spin up the
environments. Will need a careful review to make sure it meets our needs.

Fixes #50
Supersedes and Closes #47

Small update to env script while we work though details

Enforces Pyyaml be installed for the env creation script
We thought about using ruamel_yaml like Conda does, but thats only in
the base env, so we might as well enforce a more common YAML parser
if we are going to need one

Dont rely on YAML which may not be present to edit the file.

This may get scrapped in the end, but this is what it looks like

* Drop the complicated Regex-Fu in favor of more robust YAML searching
* YAML searched by PyYAML->Ruamel YAML->Conda Base Ruamel YAML through `importlib`
  * This is roboust to different conda envs, and different YAML file structures. I can enumerate everything I tried if asked
* Dropped TempFile for TempDirectory which should be more stable on Windows
* Added editing comments from reviewers
* Provided build guidelines to users on what they should chose with conditions
* Added `setup.py sdist` and `conda build` commands to the CI in the "after testing is successful" so builds happen regardless and won't block/slow testing. This ensures builds are still *part of the development process*

This last point I think finds the balance between not slowing down testing, but still including builds as part of the pipeline without the user's thought. Also avoids adding new CI builds and managing the overly exact "Allow Failures" syntax. I imagine though this may be somewhat controversial so I would like some feed back, cc @jchodera @dgasmith @loriab

Make changes to not rely on Conda-build anymore

Update the pip installs to get around pytest-cov dependency failures