[WIP] Test the GitHub Actions Workflows
This is an experimental implementation of GitHub Actions (GHA) tests.
Because we can't dry-run or construct partials on the fly for GHAs,
the route I have chosen is to have references of the current outputs
you can get from the different cookiecutter options, and have an
implementation of those outputs in the top level GHA itself.
The logic is as follows:

* Have top level GHA replicate the generated "on" block.
* Have the top level GHA generate all of the cookiecutter options which
  use GHA as a CI source. Upload all of those builds to an artifact cache.
* One job pulls the artifacts and compares the CI.yaml file outputs
  to the references, raises an error if they differ (not blocking, but
  indicates we need to change something)
  * A helper script has been included to regenerate the references.
* 3 more jobs implement the theoretical output of the cookiecutter based
  on dependency source options. This is a manual implementation but
  should do everything the live generated CI file will do. In combination
  with the reference file check from before, this should be a sufficient
  test to see if we need to make changes.

Overall, this introduces a lag-time in updating the GHA top level tests,
but I think this does successfully allow moving GHA from experimental
to live if it works.